New folder for user-contributed external terminal scripts for VM Maestro
Creating a placeholder README.md file for the folder.  We're restoring the
content that used to be on the old Discourse forum.  The core instructions
are in the CLN community at https://learningnetwork.cisco.com/docs/DOC-34612.
This folder will just be a place to collect the instructions and related
scripts for particular terminal applications.